Discover Kafka services across clusters (#2)
* Display kafka services of selected cluster in main screen

This is work in progress. Next step is to look for kafka across all clusters.

* WIP: Display all kafka services from available Kubernetes clusters

Feature works. This is quick and dirty code. Next step is to refactor code.
Reason for quick and dirty commit is CTD deadline 2022-06-20.

* Refactor code of kafka services discovery

This commit fixes lint errors, and refactors some code portions.
Next steps: polish displayed information

* Add necessary headers to Kafka Discovery Widget

* Add translations
* Get column name from translations
* Fix display of available Kafkas

* Add translation for kafkas header